+++ src/main/java/com/ll/exam/sbb/answer/repository/AnswerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer>, RepositoryUtil {
+public interface AnswerRepository extends JpaRepository<Answer, Long>, RepositoryUtil {
 
     @Transactional
     @Modifying
