+++ src/main/java/com/ll/exam/sbb/question/controller/QuestionController.java
@@ -42,7 +42,7 @@ public class QuestionController {
 
 
     @GetMapping("/detail/{id}")
-    public String detail(Model model, @PathVariable int id) {
+    public String detail(Model model, @PathVariable Long id) {
         Question question = questionService.getQuestion(id);
 
         model.addAttribute("question", question);
